[refactor] context 통합
