--- frontend/Idle/utils/carContext.jsx
@@ -1,5 +1,0 @@
-import { createContext } from "react";
-
-const carContext = createContext();
-
-export default carContext;
