+++ frontend/Idle/utils/context.jsx
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+
+export const carContext = createContext();
+export const currentPageContext = createContext();
