--- frontend/Idle/utils/currentPageContext.jsx
@@ -1,5 +1,0 @@
-import { createContext } from "react";
-
-const currentPageContext = createContext();
-
-export default currentPageContext;
